I created a reading condition for iframe or sdk,  it works.
